requirements.txt
Django==4.0.2
django-ninja==0.17.0
psycopg2-binary==2.9.3
gunicorn
